Major redesign to partly implement new filename format
Much of the code in photoname.hs has been factored into separate modules, partly so we can have more than one file naming module with good code re-use.

Added ErrorT to the Ph monad so we can deal with failure throughout the process and have access to the Options (through ReaderT) in a wider region of the code. This was necessary for some of the new switch behaviors.

Also cleaned up some of the error messages.

Adjusted existing unit tests to keep passing against new code.