Added DateFormat module for new style filename creation
This required a bit of design work on the Date module as well, adding a new function and renaming some of the existing ones to be clearer.

Unit tests were adjusted to now use --old-style, but they are painfully out of date now and I'm considering a unit test overhaul.