Display num of links via Integral vs Show
The number of links option stores the value in the same
System.Posix.Types.CNlink type as the value we get back from querying
the link count for files.

In code that displays this to the user we are now using CNlink's
Integral instance to make it into an Integer vs its Show instance.

Also removed a nearby useless use of the printf function.