End-to-end tests are now optional
Ran into a problem where the end-to-end tests fail with cabal-based CI
partly because of difficulty in the test suite finding and running the
compiled-but-uninstalled binary.
The project now defaults to running the unit tests only.

Added some arg-parsing logic to the main test module to allow a
developer to specify all tests be run via a switch like this:

    stack test --ta all

Closes gh-24.