Switched order back to DateTimeDigitized, DateTimeOriginal...
Discovered a problem with an image that was edited with Google Photos on
Android. Exif.Photo.DateTimeOriginal was modified to the edit date, no longer
reflecting when the photo was shot (which is still accurately in
Exif.Photo.DateTimeDigitized).

Not sure why this was changed in the past (in 2021, commit a6fb6906), so this
reversion will likely break something. I just don't have docs about why this
was changed in 2021.

This reverts commit a6fb6906a03f7edb5a4fbe864ede4ae4c4070d07.